+++ GAME/Actors/Player/PlayerAnimations/JumpAnimation/JumpAnimation.h
@@ -17,7 +17,7 @@ public:
 
 	// Унаследовано через IAnimation
 	virtual void flip(bool flipX, bool flipY) override;
-	virtual void setSize(int width, int height) override;
+	virtual void setSize(float width, float height) override;
 	virtual void update(float deltaTime) override;
 	virtual void draw(float posX, float posY, sf::RenderWindow & window) override;
 };
